+++ packages/mongo/lib/services.dart
@@ -45,7 +45,9 @@ Map<String, dynamic> _filterNoQuery(Map<String, dynamic> data) {
 
     if (_noQuery.contains(key) ||
         value is RequestContext ||
-        value is ResponseContext) return map;
+        value is ResponseContext) {
+      return map;
+    }
     if (key is! Map) return map..[key] = value;
     return map..[key] = _filterNoQuery(value as Map<String, dynamic>);
   });