+++ packages/orm/angel_orm_mysql/CHANGELOG.md
@@ -4,9 +4,9 @@
 
 * Require Dart >= 3.5
 * Updated `lints` to 5.0.0
-* Updated dependencies to the latest release
-* Take @SerializableField properties into account when generating `Query.parseRow` (#98)
-* Migrated test cases from `angel3_orm_test`
+* Updated dependencies to the latest releases
+* Fixed issue #98: Take `@SerializableField` properties into account when generating `Query.parseRow`
+* Migrated test cases from `angel3_orm_test` into the package
 * Removed dependency on `angel3_orm_test`
 
 ## 8.2.0