+++ packages/orm/angel_orm_postgres/CHANGELOG.md
@@ -6,8 +6,9 @@
 * Updated `lints` to 5.0.0
 * Updated dependencies to the latest release
 * Take @SerializableField properties into account when generating `Query.parseRow` (#98)
-* Migrated test cases from `angel3_orm_test`
 * Removed dependency on `angel3_orm_test`
+* Migrated test cases from `angel3_orm_test`
+* Skipped test cases for nullable primary key
 
 ## 8.2.2